Общее внесение изменений

+++ src/main/java/com/kulachok/kulachok/entity/Subscription.java
@@ -32,7 +32,7 @@ public class Subscription {
      * Идентификатор актрисы.
      */
     @Column(name = "ActrisID", nullable = false)
-    private Long actrisId;
+    private Integer actrisId;
 
     /**
      * Дата и время подписки.
